DC Bias and Heart Rate Graph
I adjusted the plate and chair rotational data by accounting for DC bias and scaling the graphs to represent degrees. I base the DC bias on the first few moments of the experiment and adjust the rest of the experiment's graph. The advantage of this is that data that is not part of an experiment is not adjusted, saving compute time.
I also fixed the heart rate graph. In its previous iteration, it did not display the proper graph. Now it gives the heart rate of the subject at every moment. This is accomplished through the use of the findpeaks() function and its "MinPeakProminence" operator in line 79.