Fixed DC Bias and Early Adjusted BP
Updated the script to adjust chair position correctly as per the test data in lines 29 and 30. Also declared a new function in line 97 onwards. The goal of this function is to adjust and plot the mid-cererbral artery as well as the heart's BP. I am not sure how to derive arterial flow, so those variables are not active. I also moved some items around to clean up the code a bit, and added the "addpath" function to try and properly allow for the code to run.